refactor(complete): Split up completion code
