+++ clap_complete/src/dynamic/mod.rs
@@ -2,8 +2,15 @@
 //!
 //! For quick-start, see [`shells::CompleteCommand`]
 
+mod candidate;
+mod complete;
 mod completer;
+mod custom;
 
 pub mod shells;
 
-pub use completer::*;
+pub use candidate::CompletionCandidate;
+pub use complete::complete;
+pub use completer::Completer;
+pub use custom::ArgValueCompleter;
+pub use custom::CustomCompleter;
